requirements.txt
streamlit
sqlalchemy
pymysql
python-dotenv
pandas
psycopg2-binary